# src/rest/v1/openapi.yml
openapi: 3.0.0
servers:
  - url: http://localhost/v1
info:
  title: ClipboardHealth
  version: 1.0.0'
components:
  schemas:
    HealthCareFacility:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
      required:
        - uuid
        - name
    HealthCareFacilityInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Worker:
      type: object
      properties:
        uuid:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        shiftAssignments:
          type: array
          items:
            $ref: '#/components/schemas/ShiftAssignment'
      required:
        - uuid
        - firstName
        - lastName
    WorkerInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
      required:
        - firstName
        - lastName
    Shift:
      type: object
      properties:
        uuid:
          type: string
        facilityUuid:
          type: string
        facility:
          $ref: '#/components/schemas/HealthCareFacility'
        shiftAssignments:
          type: array
          items:
            $ref: '#/components/schemas/ShiftAssignment'
        baseHourlyRate:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        workerSlots:
          type: integer
      required:
        - uuid
        - facilityUuid
        - baseHourlyRate
        - createdAt
        - updatedAt
        - startTime
        - endTime
        - workerSlots
...
    ShiftAssignment:
      type: object
      properties:
        shiftUuid:
          type: string
        workerUuid:
          type: string
        rating:
          type: number
        shift:
          $ref: '#/components/schemas/Shift'
        worker:
          $ref: '#/components/schemas/Worker'
      required:
        - shiftUuid
        - workerUuid
...
    RateWorkerInput:
      type: object
      properties:
        shiftUuid:
          type: string
        workerUuid:
          type: string
        rating:
          type: integer
      required:
        - shiftUuid
        - workerUuid
        - rating
paths:
...
  /facilities:
    get:
      responses:
        '200':
          description: Get HealthCareFacility list
          content:
            application/json:
              schema:
                type: object
                properties:
                  facilities:
                    type: array
                    items:
                      $ref: '#/components/schemas/HealthCareFacility'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HealthCareFacilityInput'
      responses:
        '200':
          description: Successfully created HealthCareFacility
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCareFacility'
...
  /workers:
    get:
      responses:
        '200':
          description: Get HealthCareFacility list
          content:
            application/json:
              schema:
                type: object
                properties:
                  facilities:
                    type: array
                    items:
                      $ref: '#/components/schemas/HealthCareFacility'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerInput'
      responses:
        '200':
          description: Successful health check
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
...
  /rate-worker:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateWorkerInput'
      responses:
        '200':
          description: Successfully rated a Worker for a Shift
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShiftAssignment'